ci: remove `on push`
We included `push` to make sure PRs get tested when new commits were added or the commit history was changed. However, it appears that for an active PR, every push already causes a `synchronize` activity type for the `pull_request_target` event.

So the `push` event is no longer needed for every PR. Instead, this sets it to trigger only for the case when something was pushed directly to `main`.

See:
https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#pull_request_target